requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
azure-common==1.1.26
azure-core==1.11.0
azure-mgmt-core==1.2.2
azure-mgmt-storage==17.0.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography<=3.3.2
Django==3.1.7
django-storages==1.11.1
graphene==2.1.8
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
isodate==0.6.0
jmespath==0.10.0
msrest==0.6.21
oauthlib==3.1.0
Pillow==8.1.1
promise==2.3
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.3
